This is the final version of the game
This is the final version of the game. It has been tested by my son and my nephew, 8 years  old and 6 years old, respectively.
They liked too much and started doing challenges.
There are new steps, such as create an executable (to avoid to require Python installation) and maybe insert different types of pandemy.
But for a while it is an stable version of my development