Changing the script to compare two texts
The final course project had as input a text signature. This, version I've changed to receive a text as input and to calculate the signature of the reference text. The intention is to return the similarity of the signatures and leave the user to decide if the texts could be similar enough to be considered a plagiarism